pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ipa-core"
version = "0.0.0"
description = "Reconocimiento y evaluación de pronunciación en IPA"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn>=0.24,<1",
    "python-multipart>=0.0.6,<0.0.10",
    "PyYAML>=6.0,<7",
    "typer>=0.9.0,<1",
    "rich>=13.0.0,<14",
    "numpy>=1.26,<3",
]

[project.optional-dependencies]
asr = [
    "allosaurus>=1.0.0,<2",
]
ollama = [
    "ollama>=0.1,<1",
]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=4.1,<6",
    "httpx>=0.27,<0.28",
    "mypy>=1.11,<2",
]
speech = [
    "sounddevice>=0.4.6,<1",
    "pydub>=0.25.1,<1",
    "epitran>=1.24,<2",
]
onnx = [
    "onnxruntime>=1.16,<2",
    "librosa>=0.10,<1",
    "soundfile>=0.12,<1",
    "numpy>=1.26,<3",
]

[project.scripts]
pronunciapa = "ipa_core.interfaces.cli:main"

[tool.setuptools]
packages = ["ipa_core", "ipa_plugin_allosaurus"]
package-dir = {"" = ".", "ipa_plugin_allosaurus" = "plugins/allosaurus/ipa_plugin_allosaurus"}

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["ipa_core"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[project.entry-points."pronunciapa.plugins"]
"asr.stub_ep" = "ipa_core.backends.asr_stub:StubASR"
"asr.allosaurus" = "ipa_plugin_allosaurus.backend:AllosaurusASR"
"asr.onnx" = "ipa_core.plugins.asr_onnx:ONNXASRPlugin"
"asr.whisper_onnx" = "ipa_core.plugins.asr_onnx:ONNXASRPlugin"
"asr.whisper" = "ipa_core.plugins.asr_onnx:ONNXASRPlugin"
"tts.default" = "ipa_core.tts.adapter:TTSAdapter"
"tts.piper" = "ipa_core.tts.piper:PiperTTS"
"tts.system" = "ipa_core.tts.system:SystemTTS"
"llm.llama_cpp" = "ipa_core.llm.llama_cpp:LlamaCppAdapter"
"llm.onnx" = "ipa_core.llm.onnx:OnnxLLMAdapter"
"llm.stub" = "ipa_core.llm.stub:StubLLMAdapter"

